image added
we've added image in our front page. Here we created a widget called ImageBanner.
ImageBanner take the path of an image as it's parameter. Then we set the image in
a container of height 200 px. We set the background color grey and then set the
BoxFit.cover to cover whole div.

We imported the image_banner.dart into the location_detail.dart file. Then we set
the ImageBanner widget into the Column widget.