location model implementation
Now we implemented location model to our location detail widget
we called fetchAll() method to get all locations. As we have
only one location for now we get its first location. Then,
location.name inserted into the title section, imagePath into the
ImageBanner widget and textSection of location cascaded with the children.

textSection() get the location and converted it to the list of widget.